Move Lookup to Index
