+++ src/Curry/LanguageServer/Handlers/Hover.hs
@@ -8,7 +8,7 @@ import Control.Monad.Trans (liftIO)
 import Control.Monad.Trans.Maybe
 import qualified Curry.LanguageServer.Index.Store as I
 import Curry.LanguageServer.Utils.Convert (ppPredTypeToText, currySpanInfo2Range)
-import Curry.LanguageServer.Utils.Lookup
+import Curry.LanguageServer.Index.Lookup
 import Curry.LanguageServer.Utils.General (liftMaybe)
 import Curry.LanguageServer.Utils.Syntax (TypedSpanInfo (..), ModuleAST, moduleIdentifier)
 import Curry.LanguageServer.Utils.Uri (normalizeUriWithPath)
